requirements.txt
asgiref==3.6.0
Django==4.1.5
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
django-extensions==3.2.1
python-environ==0.4.54

# render requirements
brotli==1.1.0
whitenoise==6.9.0
uvicorn==0.34.0